mods.py
# System imports
import sys
import os
import subprocess
import re
import random
import time

# File management imports
from glob import glob
from natsort import natsorted
import shutil
import string

# Wikipedia
import wikipedia
 
# NLP imports
import nltk.data
#To be improved

# Stable Whisper (Subtitles)
import stable_whisper
import pysrt

# Image Management
from cropimage import Cropper
import cv2
import ssl
ssl._create_default_https_context = ssl._create_unverified_context
from PIL import Image
import open_clip
from icrawler.builtin import BingImageCrawler

# PyTorch
import torch

# Audio
import librosa
#To be improved

# Flask
from flask import Flask, request

# Article Summarization
from transformers import AutoTokenizer, AutoModelForSeq2SeqLM
from newspaper import Article

# OpenVoice Local Files
from se_extractor import get_se
from api import BaseSpeakerTTS, ToneColorConverter

# Export all these for use in other files

__all__ = ['sys', 'os', 'subprocess', 're', 'random', 'time', 'glob', 'natsorted', 'shutil', 'string', 'wikipedia', 'nltk', 'stable_whisper', 'pysrt', 'Cropper', 'cv2', 'ssl', 'Image', 'open_clip', 'BingImageCrawler', 'torch', 'librosa', 'AutoTokenizer', 'AutoModelForSeq2SeqLM', 'Article', 'BaseSpeakerTTS', 'ToneColorConverter', 'get_se', 'Flask', 'request']
