requirements.txt
natsort
wikipedia
nltk
stable-ts
pysrt
cropimage
#opencv-python-headless
opencv-contrib-python
Pillow
open-clip-torch
icrawler
torch
librosa
transformers
newspaper3k
flask
facenet-pytorch
pydub
faster-whisper
whisper-timestamped
dtw-python
inflect
unidecode
eng_to_ipa
wavmark

